+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - GLS-124 - new homepage banner
 - NOTICKET - retire Enzyme and cleanup tests output
 - GLS-152 - Add CTA banner to homepage
+- GLS-180 - Add trade visualisations to Country guide pages
 
 ## [2.6.0](https://github.com/uktrade/great-cms/releases/tag/2.6.0)